requirements.txt
# Common utilities
requests
ipython

# Djando + util
django
pytils
Pillow
sorl-thumbnail
python-dotenv

# API
djoser
djangorestframework-simplejwt
djangorestframework
PyJWT

# Debagger
django-debug-toolbar

# Distributed async tasks
celery
redis
django-celery-beat

# Django async requirements
gunicorn
uvicorn

# lintering
flake8
flake8-broken-line
flake8-isort
flake8-plugin-utils
flake8-return
pep8-naming

# Bot & utils
python-telegram-bot
BeautifulSoup4
python-dateutil
dateparser
tzlocal
tzwhere